move sheet fetching logic from layout to sheet page closes #117
