+++ client/src/pages/layout.vue
@@ -11,7 +11,7 @@
 </template>
 
 <script>
-const { mapState, mapActions, mapMutations } = require('vuex')
+const { mapState, mapMutations } = require('vuex')
 const values = require('lodash/values')
 
 const SiteNav = require('../components/site-nav.vue')
@@ -20,43 +20,17 @@ const Notification = require('../components/notification.vue')
 module.exports = {
   name: 'Layout',
   computed: mapState({
-    sheets: (state) => state.db.sheets,
     notifications: (state) => values(state.ui.notifications)
   }),
-  methods: {
-    ...mapActions([
-      'getSheetList',
-      'getSheet'
-    ]),
-    ...mapMutations([
-      'dismissNotification'
-    ]),
-    getActiveSheet () {
-      const route = this.$route.params.sheetName
-      const activeSheetName = determineActiveSheet(route, this.sheets)
-      if (activeSheetName) this.getSheet({ name: activeSheetName })
-    }
-  },
-  async created () {
-    await this.getSheetList()
-    this.getActiveSheet()
-  },
-  watch: {
-    '$route': 'getActiveSheet'
-  },
+  methods: mapMutations([
+    'dismissNotification'
+  ]),
   components: {
     'site-nav': SiteNav,
     'notification': Notification
   }
 }
 
-function determineActiveSheet (route, sheets) {
-  if (route) {
-    return route
-  } else if (sheets.length > 0) {
-    return sheets[0].name
-  }
-}
 </script>
 
 <style src="../../../node_modules/bulma/css/bulma.css"></style>
