+++ client/src/pages/sheet.vue
@@ -16,7 +16,7 @@
 </template>
 
 <script>
-const { mapActions } = require('vuex')
+const { mapActions, mapState } = require('vuex')
 
 const SiteNav = require('../components/site-nav.vue')
 const SheetList = require('../components/sheet-list.vue')
@@ -31,9 +31,36 @@ module.exports = {
     'sheet-name': SheetName,
     'grid': Grid
   },
-  methods: mapActions([
-    'insertSheet'
-  ])
+  computed: mapState({
+    sheets: (state) => state.db.sheets,
+  }),
+  methods: {
+    ...mapActions([
+      'insertSheet',
+      'getSheetList',
+      'getSheet'
+    ]),
+    getActiveSheet () {
+      const route = this.$route.params.sheetName
+      const activeSheetName = determineActiveSheet(route, this.sheets)
+      if (activeSheetName) this.getSheet({ name: activeSheetName })
+    }
+  },
+  async created () {
+    await this.getSheetList()
+    this.getActiveSheet()
+  },
+  watch: {
+    '$route': 'getActiveSheet'
+  },
+}
+
+function determineActiveSheet (route, sheets) {
+  if (route) {
+    return route
+  } else if (sheets.length > 0) {
+    return sheets[0].name
+  }
 }
 </script>
